add value check in gtest

+++ gtest/src/sample.rs
@@ -46,7 +46,7 @@ pub struct Program {
     pub id: u64,
     pub init_message: Option<PayloadVariant>,
     pub init_gas_limit: Option<u64>,
-    pub init_value: Option<u64>,
+    pub init_value: Option<u128>,
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -151,7 +151,7 @@ pub struct Message {
     pub destination: u64,
     pub payload: Option<PayloadVariant>,
     pub gas_limit: Option<u64>,
-    pub value: Option<u64>,
+    pub value: Option<u128>,
     #[serde(rename = "exitCode")]
     pub exit_code: Option<i32>,
 }
